[refactor] 불필요한 코드 삭제 및 UserService 디렉토리 구조 수정
