--- src/main/java/qp/official/qp/converter/ImageConverter.java
@@ -1,5 +1,0 @@
-package qp.official.qp.converter;
-
-public class ImageConverter {
-
-}
