+++ src/main/java/qp/official/qp/service/AnswerService/AnswerCommandService.java
@@ -2,15 +2,13 @@ package qp.official.qp.service.AnswerService;
 
 import qp.official.qp.domain.Answer;
 import qp.official.qp.domain.enums.AnswerLikeStatus;
-import qp.official.qp.domain.mapping.AnswerLikes;
-import qp.official.qp.web.dto.AnswerLikeRequestDTO;
 import qp.official.qp.web.dto.AnswerRequestDTO;
 
 public interface AnswerCommandService {
-    public Answer createAnswer(AnswerRequestDTO.AnswerCreateDTO request, Long questionId);
+     Answer createAnswer(AnswerRequestDTO.AnswerCreateDTO request, Long questionId);
 
 
-    public AnswerLikeStatus addAndDeleteLikeToAnswer(Long userId, Long answerId);
+     AnswerLikeStatus addAndDeleteLikeToAnswer(Long userId, Long answerId);
 
 
     void deleteAnswer(Long answerId);
