+++ src/main/java/qp/official/qp/service/TokenService/TokenService.java
@@ -5,6 +5,5 @@ import qp.official.qp.web.dto.TokenResponseDTO;
 public interface TokenService {
     TokenResponseDTO createToken(Long userId);
     TokenResponseDTO isValidToken(Long userId);
-
     TokenResponseDTO autoSignIn(Long userId);
 }
