+++ src/main/java/qp/official/qp/service/UserService/UserService.java
@@ -1,18 +1,15 @@
-package qp.official.qp.service;
+package qp.official.qp.service.UserService;
 
 import org.json.simple.parser.ParseException;
 import qp.official.qp.domain.User;
 import qp.official.qp.web.dto.UserRequestDTO;
-import qp.official.qp.web.dto.UserResponseDTO;
 import java.io.IOException;
-import java.util.Map;
 public interface UserService {
 
     User getUserInfo(Long userId);
     User updateUserInfo(Long userId, UserRequestDTO.UpdateUserInfoRequestDTO requestDTO);
     User createTestUser();
     User signUp(String code) throws IOException, ParseException;
-
     User autoSignIn(Long userId);
 
 }
