+++ src/main/java/qp/official/qp/web/controller/UserRestController.java
@@ -12,7 +12,7 @@ import qp.official.qp.apiPayload.code.status.SuccessStatus;
 import qp.official.qp.converter.UserConverter;
 import qp.official.qp.domain.User;
 import qp.official.qp.service.TokenService.TokenService;
-import qp.official.qp.service.UserService;
+import qp.official.qp.service.UserService.UserService;
 import qp.official.qp.validation.annotation.ExistUser;
 import qp.official.qp.web.dto.TokenResponseDTO;
 import qp.official.qp.web.dto.UserRequestDTO;
