--- src/main/java/qp/official/qp/web/dto/AnswerLikeRequestDTO.java
@@ -1,9 +1,0 @@
-package qp.official.qp.web.dto;
-
-import lombok.Getter;
-import qp.official.qp.validation.annotation.ExistAnswer;
-import qp.official.qp.validation.annotation.ExistUser;
-
-public class AnswerLikeRequestDTO {
-
-}
