+++ src/main/java/qp/official/qp/web/dto/AnswerResponseDTO.java
@@ -88,22 +88,4 @@ public class AnswerResponseDTO {
         LocalDateTime updatedAt;
     }
 
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class GetAnswerResultDTO{
-        Long answerId;
-        String title;
-        String content;
-        String category;
-        Long userId;
-        Long question_id;
-        Long likeCount;
-        boolean isParent;
-        List<GetAnswerResultDTO> children;
-        LocalDateTime createdAt;
-        LocalDateTime updatedAt;
-    }
-
 }
