--- src/main/java/qp/official/qp/web/dto/TokenRequestDTO.java
@@ -1,15 +1,0 @@
-package qp.official.qp.web.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenRequestDTO {
-    private String accessToken;
-    private String refreshToken;
-}
